refactor(ecmwf): move the retry helpers into _helpers.py
CLAUDE.md reserves backend.py for the AbstractDataSource subclass and
_helpers.py for private stateless helpers, and every sibling backend
facing a flaky upstream keeps its typed error and classifiers there -
gdacs, osm, bathymetry. Seven new module-level symbols had gone into
backend.py instead, taking it past 1800 lines.

Move the error, both classifiers, the status extractor, the retry
constants and the retry loop into a new ecmwf/_helpers.py; backend.py
drops to ~1580.

L10 - on a budget overrun the fixture failed while the daemon thread was
still writing into tmp_path, which pytest then removed underneath it.
Give the thread a short grace period and say so in the failure when it
outlives that, rather than leaving a teardown error as the only clue.

N5 - _variables_in and _time_stamps each re-implemented the unzip into
different directories, extracting a zipped result twice, and an archive
with no NetCDF member surfaced as an empty variable set or a bare
IndexError. One _netcdf_members does it once and says what went wrong.
